src/requirements.txt
aiogram==3.13.1
requests==2.31.0
python-dotenv==1.0.1
fastapi==0.115.2
uvicorn==0.31.1
pydantic==2.9.2
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
aiosqlite